Traverse disc nodes
Especially when a discretionary is given explicitly or contains
ligatures, the inner list might have a different color attribute than
the disc node. So traverse the replace element like any other nested
list instead of treating a disc as an atomic colored object.

We only have to look at replace because if TeX actually breaks the line
at the disc and .pre/.post is used, TeX moves the lists out of the
discretionary. Then the disc only contains empty sublists.